- Fixed std::min/max issues on linux gcc - Fixed missing fstream include for linux gcc - A few apps require boost 1.50.0 instead of 1.48.0 (for fs::path operator+)

+++ libRender/src/render/utils.cpp
@@ -40,10 +40,10 @@ cv::Rect calculateBoundingBox(Vertex v0, Vertex v1, Vertex v2, int viewportWidth
 	t.minY = max(min(t.v0.position[1], min(t.v1.position[1], t.v2.position[1])), 0.0f);
 	t.maxY = min(max(t.v0.position[1], max(t.v1.position[1], t.v2.position[1])), (float)(viewportHeight - 1));*/
 	
-	int minX = max(min(floor(v0.position[0]), min(floor(v1.position[0]), floor(v2.position[0]))), 0.0f);
-	int maxX = min(max(ceil(v0.position[0]), max(ceil(v1.position[0]), ceil(v2.position[0]))), static_cast<float>(viewportWidth - 1));
-	int minY = max(min(floor(v0.position[1]), min(floor(v1.position[1]), floor(v2.position[1]))), 0.0f);
-	int maxY = min(max(ceil(v0.position[1]), max(ceil(v1.position[1]), ceil(v2.position[1]))), static_cast<float>(viewportHeight - 1));
+    int minX = max(min(floor(v0.position[0]), min(floor(v1.position[0]), floor(v2.position[0]))), 0.0); // What about rounding, or rather the conversion from double to int?
+    int maxX = min(max(ceil(v0.position[0]), max(ceil(v1.position[0]), ceil(v2.position[0]))), static_cast<double>(viewportWidth - 1)); // and why can't we do this with float in linux?
+    int minY = max(min(floor(v0.position[1]), min(floor(v1.position[1]), floor(v2.position[1]))), 0.0);
+    int maxY = min(max(ceil(v0.position[1]), max(ceil(v1.position[1]), ceil(v2.position[1]))), static_cast<double>(viewportHeight - 1));
 	return cv::Rect(minX, minY, maxX - minX, maxY - minY);
 }
